requirements.txt
pandas
pytest
streamlit
anthropic>=0.40.0
python-dotenv>=1.0.0
